Better handling if network connectivity lost
In the event of a network error, a helper method that checks the socket data returned by the MySQL server to check the packet type could be null and would therefore throw a NullPointerException.
Instead, now a MySQLConnException will be thrown so it will be returned back with the callback to the database request.